pyproject.toml
[project]
name = "manimproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.8",
    "langchain-openai>=1.0.3",
    "langchain-google-genai>=2.0.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
]
